+++ src/newgrf_station.h
@@ -45,7 +45,7 @@ struct StationScopeResolver : public ScopeResolver {
 	uint32 GetRandomBits() const override;
 	uint32 GetTriggers() const override;
 
-	uint32 GetVariable(byte variable, uint32 parameter, GetVariableExtra *extra) const override;
+	uint32 GetVariable(uint16 variable, uint32 parameter, GetVariableExtra *extra) const override;
 };
 
 /** Station resolver. */
